+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -12,6 +12,7 @@ major version release (i.e. 1.0.0).
 
 * Added native support for Internet Explorer - the Java Selenium server is no
     longer required.
+* Added native support for Safari
 * Added native support for Opera 26+
 * With the release of [Node 0.12.0](http://blog.nodejs.org/2015/02/06/node-v0-12-0-stable/)
     (finally!), the minimum supported version of Node is now `0.10.x`.